Reorganize assessment components
