+++ src/views/assignments/AssignmentsOverview.vue
@@ -84,9 +84,9 @@ import 'confetti-js';
 
 // Page components
 import AssignmentOverviewStats from '@/views/components/assignments/overview/AssignmentOverviewStats';
-import AssessmentOverviewDescription from '@/views/components/AssessmentOverviewDescription';
-import AssessmentOverviewActionButtons from '@/views/components/AssessmentOverviewActionButtons';
-import AssessmentOverviewTitle from '@/views/components/AssessmentOverviewTitle';
+import AssessmentOverviewDescription from '@/views/components/assessment/AssessmentOverviewDescription';
+import AssessmentOverviewActionButtons from '@/views/components/assessment/AssessmentOverviewActionButtons';
+import AssessmentOverviewTitle from '@/views/components/assessment/AssessmentOverviewTitle';
 import AssignmentsModalEdit from '@/views/components/assignments/AssignmentsModalEdit';
 import AssignmentOverviewTabs from '@/views/components/assignments/overview/AssignmentOverviewTabs';
