+++ src/views/components/assessment/AssessmentOverviewActionButtons.vue
